requirements.txt
argbind>=0.3.7
descript-audiotools>=0.7.2
einops
numpy
torchaudio
tqdm
tensorboard
numba>=0.5.7
jupyterlab
descript-audio-codec

